Refactor: Token 폴더 따로 만들어서 모아두기
