+++ src/main/java/com/kwcapstone/Token/Service/TokenService.java
@@ -1,9 +1,13 @@
-package com.kwcapstone.Kakao;
+package com.kwcapstone.Token.Service;
 
 import com.kwcapstone.Repository.MemberRepository;
+import com.kwcapstone.Token.Domain.Token;
+import com.kwcapstone.Token.Domain.Dto.TokenResponse;
+import com.kwcapstone.Token.JwtTokenProvider;
+import com.kwcapstone.Token.Domain.Convert.TokenConvert;
+import com.kwcapstone.Token.Repository.TokenRepository;
 import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
-import org.bson.types.ObjectId;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
